refactor : 가게 상세 정보 API를 위한 Dto 패키지 분리
