+++ data/src/main/java/com/example/data/dto/response/store/CoordinateDto.kt
@@ -1,9 +1,8 @@
-package com.example.data.dto.response
+package com.example.data.dto.response.store
 
 import com.example.domain.model.map.CoordinateModel
 import kotlinx.serialization.Serializable
 
-@Serializable
 data class CoordinateDto(
     val longitude: Double,
     val latitude: Double,
