+++ data/src/main/java/com/example/data/dto/response/store/OpeningHoursDto.kt
@@ -1,9 +1,7 @@
-package com.example.data.dto.response
+package com.example.data.dto.response.store
 
 import com.example.domain.model.map.OpeningHoursModel
-import kotlinx.serialization.Serializable
 
-@Serializable
 data class OpeningHoursDto(
     val open: TimeInfoData,
     val close: TimeInfoData,
