+++ data/src/main/java/com/example/data/dto/response/store/TimeInfoData.kt
@@ -1,9 +1,8 @@
-package com.example.data.dto.response
+package com.example.data.dto.response.store
 
 import com.example.domain.model.map.TimeInfoModel
 import kotlinx.serialization.Serializable
 
-@Serializable
 data class TimeInfoData(
     val day: String,
     val hour: Int,
